Move functions into Nxmap as class objects.
